services/sensor-processor/requirements.txt
fastapi
uvicorn[standard]
kafka-python
prometheus-client
python-dotenv